update posts according to the number of likes
Like a post. Note: In the MiniWall a post owner cannot comment or like her/his post. The MiniWall always show posts according to the number of likes. Popular posts, with high number of likes are on the top. Posts with the same number of likes are shown in reverse chronological order of the date of post.